+++ cockatrice/src/game/board/card_item.h
@@ -3,7 +3,7 @@
 
 #include "../zones/logic/card_zone_logic.h"
 #include "abstract_card_item.h"
-#include "server_card.h"
+#include "server/game/server_card.h"
 
 class CardDatabase;
 class CardDragItem;
