+++ cockatrice/src/server/local_server.cpp
@@ -1,7 +1,7 @@
 #include "local_server.h"
 
 #include "local_server_interface.h"
-#include "server_room.h"
+#include "server/server_room.h"
 
 LocalServer::LocalServer(QObject *parent) : Server(parent)
 {
