+++ cockatrice/src/server/local_server_interface.h
@@ -1,7 +1,7 @@
 #ifndef LOCALSERVERINTERFACE_H
 #define LOCALSERVERINTERFACE_H
 
-#include "server_protocolhandler.h"
+#include "server/server_protocolhandler.h"
 
 class LocalServer;
