+++ common/server/game/server_game.h
@@ -20,10 +20,10 @@
 #ifndef SERVERGAME_H
 #define SERVERGAME_H
 
+#include "../server_response_containers.h"
 #include "pb/event_leave.pb.h"
 #include "pb/response.pb.h"
 #include "pb/serverinfo_game.pb.h"
-#include "server_response_containers.h"
 
 #include <QDateTime>
 #include <QMap>
