+++ common/server/game/server_player.cpp
@@ -1,9 +1,15 @@
 #include "server_player.h"
 
-#include "color.h"
-#include "deck_list.h"
-#include "deck_list_card_node.h"
-#include "get_pb_extension.h"
+#include "../../color.h"
+#include "../../deck_list.h"
+#include "../../deck_list_card_node.h"
+#include "../../get_pb_extension.h"
+#include "../../rng_abstract.h"
+#include "../../trice_limits.h"
+#include "../server.h"
+#include "../server_abstractuserinterface.h"
+#include "../server_database_interface.h"
+#include "../server_room.h"
 #include "pb/command_attach_card.pb.h"
 #include "pb/command_change_zone_properties.pb.h"
 #include "pb/command_concede.pb.h"
@@ -70,17 +76,11 @@
 #include "pb/response_dump_zone.pb.h"
 #include "pb/serverinfo_player.pb.h"
 #include "pb/serverinfo_user.pb.h"
-#include "rng_abstract.h"
-#include "server.h"
-#include "server_abstractuserinterface.h"
 #include "server_arrow.h"
 #include "server_card.h"
 #include "server_cardzone.h"
 #include "server_counter.h"
-#include "server_database_interface.h"
 #include "server_game.h"
-#include "server_room.h"
-#include "trice_limits.h"
 
 #include <QDebug>
 #include <QRegularExpression>
