+++ common/server/server.cpp
@@ -19,18 +19,17 @@
  ***************************************************************************/
 #include "server.h"
 
-#include "debug_pb_message.h"
-#include "featureset.h"
+#include "../debug_pb_message.h"
+#include "../featureset.h"
+#include "game/server_game.h"
+#include "game/server_player.h"
 #include "pb/event_connection_closed.pb.h"
 #include "pb/event_list_rooms.pb.h"
 #include "pb/event_user_joined.pb.h"
 #include "pb/event_user_left.pb.h"
 #include "pb/isl_message.pb.h"
 #include "pb/session_event.pb.h"
-#include "server_counter.h"
 #include "server_database_interface.h"
-#include "server_game.h"
-#include "server_player.h"
 #include "server_protocolhandler.h"
 #include "server_remoteuserinterface.h"
 #include "server_room.h"
