+++ common/server/server_room.cpp
@@ -1,5 +1,7 @@
 #include "server_room.h"
 
+#include "../trice_limits.h"
+#include "game/server_game.h"
 #include "pb/commands.pb.h"
 #include "pb/event_join_room.pb.h"
 #include "pb/event_leave_room.pb.h"
@@ -9,9 +11,7 @@
 #include "pb/room_commands.pb.h"
 #include "pb/serverinfo_chat_message.pb.h"
 #include "pb/serverinfo_room.pb.h"
-#include "server_game.h"
 #include "server_protocolhandler.h"
-#include "trice_limits.h"
 
 #include <QDateTime>
 #include <QDebug>
