+++ servatrice/src/servatrice.cpp
@@ -29,8 +29,8 @@
 #include "pb/event_server_shutdown.pb.h"
 #include "servatrice_connection_pool.h"
 #include "servatrice_database_interface.h"
+#include "server/server_room.h"
 #include "server_logger.h"
-#include "server_room.h"
 #include "serversocketinterface.h"
 #include "settingscache.h"
 #include "smtpclient.h"
