+++ servatrice/src/serversocketinterface.cpp
@@ -65,10 +65,10 @@
 #include "pb/serverinfo_user.pb.h"
 #include "servatrice.h"
 #include "servatrice_database_interface.h"
+#include "server/game/server_player.h"
+#include "server/server_response_containers.h"
+#include "server/server_room.h"
 #include "server_logger.h"
-#include "server_player.h"
-#include "server_response_containers.h"
-#include "server_room.h"
 #include "settingscache.h"
 #include "trice_limits.h"
 #include "version_string.h"
